==> 04-VotingProfile/Clustering/5-Link_maker.ipynb <==
Load matrx distance & find 3 closest nbh

import numpy as np
import pandas as pd

df.head()

names = df['Unnamed: 0']
df.drop('Unnamed: 0',axis=1,inplace=True)
df.head()

df.head()

dic_ = {'source':[],"target":[], "value":[]}
matrix_dim = df.shape
print(matrix_dim)
columns = df.columns
for index in range(matrix_dim[0]) :
    for column in range(index,matrix_dim[1]) :
        if df[columns[column]][index] !=0 :
            dic_['source'] +=[names[index]]
            dic_['target'] +=[columns[column]]
            dic_['value'] +=[df[columns[column]][index]]

frame = pd.DataFrame(dic_)
frame.shape

frame.to_json('linksSubject.json')
